Uncomment the call to channel/2 in HelloPhoenix.UserSocket
In `web/channels/user_socket.ex`, the `HelloPhoenix.UserSocket` we pointed to in our endpoint has already been created when we generated our application. We need to make sure messages get routed to the correct channel. To do that, we'll uncomment the "rooms:*" channel definition:

    # web/channels/user_socket.ex
    defmodule HelloPhoenix.UserSocket do
      use Phoenix.Socket

      ## Channels
      channel "rooms:*", HelloPhoenix.RoomChannel
      ...
    end

Now, whenever a client sends a message whose topic starts with `"rooms:"`, it will be routed to our RoomChannel. Next, we'll define a `HelloPhoenix.RoomChannel` module to manage our chat room messages.