Add `HelloPhoenix.RoomChannel`
### Joining Channels

The first priority of your channels is to authorize clients to join a given topic. For authorization, we must implement `join/3` in `web/channels/room_channel.ex`.

    # web/channels/room_channel.ex
    defmodule HelloPhoenix.RoomChannel do
      use Phoenix.Channel

      def join("rooms:lobby", _message, socket) do
        {:ok, socket}
      end
      def join("rooms:" <> _private_room_id, _params, _socket) do
        {:error, %{reason: "unauthorized"}}
      end
    end

For our chat app, we'll allow anyone to join the `"rooms:lobby"` topic, but any other room will be considered private and special authorization, say from a database, will be required. We won't worry about private chat rooms for this exercise, but feel free to explore after we finish. To authorize the socket to join a topic, we return `{:ok, socket}` or `{:ok, reply, socket}`. To deny access, we return `{:error, reply}`. More information about authorization with tokens can be found in the [`Phoenix.Token` documentation](http://hexdocs.pm/phoenix/Phoenix.Token.html).