Add a "new_msg"-listener to socket.js
All we had to do is detect that enter was pressed and then `push` an event over the channel with the message body. We named the event "new_msg". With this in place, let's handle the other piece of a chat application where we listen for new messages and append them to our messages container.

    # web/static/js/socket.js
    ...
    let channel           = socket.channel("rooms:lobby", {})
    let chatInput         = $("#chat-input")
    let messagesContainer = $("#messages")

    chatInput.on("keypress", event => {
      if(event.keyCode === 13){
        channel.push("new_msg", {body: chatInput.val()})
        chatInput.val("")
      }
    })

    channel.on("new_msg", payload => {
      messagesContainer.append(`<br/>[${Date()}] ${payload.body}`)
    })

    channel.join()
      .receive("ok", resp => { console.log("Joined successfully", resp) })
      .receive("error", resp => { console.log("Unable to join", resp) })

    export default socket

We listen for the `"new_msg"` event using `channel.on`, and then append the message body to the DOM. Now let's handle the incoming and outgoing events on the server to complete the picture.